Retain the original commit message
Because I want to keep the original commit message intact, make
a small hack for now to retain it when a message is not passed
via CLI. Will later try to merge the two so it retains both the
original message and the update message, but for now this does
what I want in the short term.

This keeps with my view that the one thing a tool should not do
is erase work that was already done.